requirements.txt
fastapi
uvicorn
pydantic
requests
playwright
python-dotenv
openai
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
transformers
python-mecab-ko
numpy
xgboost==1.7.6
shap
paddleocr